+++ spec/dummy/app/assets/javascripts/dashing/widgets/index.js
@@ -10,7 +10,4 @@
 // WARNING: THE FIRST BLANK LINE MARKS THE END OF WHAT'S TO BE PROCESSED, ANY BLANK LINE SHOULD
 // GO AFTER THE REQUIRES BELOW.
 //
-//= require dashing
 //= require_tree .
-//= require dashing.gridster
-//= require widgets