+++ spec/lib/dashing/configuration_spec.rb
@@ -4,22 +4,32 @@ describe Dashing::Configuration do
 
   let(:instance) { Dashing::Configuration.new }
 
-  it { expect(instance.scheduler).to              be_a(::Rufus::Scheduler) }
+  it { expect(instance.engine_path).to            eq('/dashing') }
+  # it { expect(instance.scheduler).to              be_a(::Rufus::Scheduler.new) }
+  it { expect(instance.redis).to                  be_a(::Redis) }
+
+  # Redis
   it { expect(instance.redis_host).to             eq('127.0.0.1') }
   it { expect(instance.redis_port).to             eq('6379') }
   it { expect(instance.redis_password).to         be_nil }
   it { expect(instance.redis_namespace).to        eq('dashing_events') }
-  it { expect(instance.view_path).to              eq('app/views/dashing/') }
-  it { expect(instance.jobs_path).to              eq('app/jobs/') }
-  it { expect(instance.engine_path).to            eq('/dashing') }
-  it { expect(instance.dashboards_path).to        eq('app/views/dashing/dashboards/') }
-  it { expect(instance.dashboard_layout).to       eq('dashing/dashboard') }
-  it { expect(instance.widgets_path).to           eq('app/views/dashing/widgets/') }
-  it { expect(instance.default_dashboard).to      be_nil }
+
+  # Authorization
   it { expect(instance.auth_token).to             be_nil }
   it { expect(instance.devise_allowed_models).to  be_empty }
 
-  it { expect(instance.redis).to                  be_a(::Redis) }
+  # Jobs
+  it { expect(instance.jobs_path).to              eq('app/jobs/') }
+
+  # Dashboards
+  it { expect(instance.default_dashboard).to      be_nil }
+  it { expect(instance.dashboards_views_path).to  eq('app/views/dashing/dashboards/') }
+  it { expect(instance.dashboard_layout_path).to  eq('dashing/dashboard') }
+
+  # Widgets
+  it { expect(instance.widgets_views_path).to     eq('app/views/dashing/widgets/') }
+  it { expect(instance.widgets_js_path).to        eq('app/assets/javascripts/dashing') }
+  it { expect(instance.widgets_css_path).to       eq('app/assets/stylesheets/dashing') }
 
   describe '#request_thread_count' do